Alterações e correções nas configurações do projeto

+++ GestaoEntregasDAO/src/main/java/br/cefetmg/gestaoentregasdao/dao/DAO.java
@@ -41,7 +41,7 @@ public class DAO<T> {
         }
     }
 
-    public T consultar(Long id) throws DAOException {
+    public T consultar(Integer id) throws DAOException {
         EntityManager entityManager = null;
         T result;
         try {
